+++ components/forms/Checkbox/stories.html.js
@@ -3,4 +3,4 @@ import { action } from "@storybook/addon-actions";
 import html from "./html.html";
 import style from "./style.module";
 
-storiesOf("Internal/Forms/Checkbox", module).add("default", () => html);
+storiesOf("Forms/Checkbox", module).add("default", () => html);