+++ components/forms/Input/stories.react.js
@@ -4,7 +4,7 @@ import { withKnobs, boolean, select } from "@storybook/addon-knobs";
 import { withState } from "@dump247/storybook-state";
 import Input from "./react";
 
-storiesOf("Internal/Forms/Input", module)
+storiesOf("Forms/Input", module)
   .addDecorator(withKnobs)
   .add("default", () => {
     const variant = select(